requirements.txt
python-fasthtml==0.4.4
sqlite_minutils==4.0.3
uvicorn==0.30.1
requests
